+++ examples/app_demo.rs
@@ -5,7 +5,7 @@
 //! rust_xlsxwriter library.
 
 use chrono::NaiveDate;
-use rust_xlsxwriter::{Format, Workbook, XlsxError};
+use rust_xlsxwriter::{Format, Workbook, XlsxAlign, XlsxBorder, XlsxError};
 
 fn main() -> Result<(), XlsxError> {
     // Create a new Excel file object.
@@ -15,6 +15,9 @@ fn main() -> Result<(), XlsxError> {
     let bold_format = Format::new().set_bold();
     let decimal_format = Format::new().set_num_format("0.000");
     let date_format = Format::new().set_num_format("yyyy-mm-dd");
+    let merge_format = Format::new()
+        .set_border(XlsxBorder::Thin)
+        .set_align(XlsxAlign::Center);
 
     // Add a worksheet to the workbook.
     let worksheet = workbook.add_worksheet();
@@ -38,9 +41,12 @@ fn main() -> Result<(), XlsxError> {
     // Write a formula.
     worksheet.write_formula_only(5, 0, "=SIN(PI()/4)")?;
 
-    // Write the date .
+    // Write a date.
     let date = NaiveDate::from_ymd(2023, 1, 25);
     worksheet.write_date(6, 0, date, &date_format)?;
+
+    // Write some merged cells.
+    worksheet.merge_range(7, 0, 7, 1, "Merged cells", &merge_format)?;
 
     // Save the file to disk.
     workbook.save("demo.xlsx")?;