+++ src/workbook.rs
@@ -31,7 +31,7 @@ use crate::{XlsxColor, XlsxPattern};
 /// # // This code is available in examples/app_demo.rs
 /// #
 /// use chrono::NaiveDate;
-/// use rust_xlsxwriter::{Format, Workbook, XlsxError};
+/// use rust_xlsxwriter::{Format, Workbook, XlsxAlign, XlsxBorder, XlsxError};
 ///
 /// fn main() -> Result<(), XlsxError> {
 ///     // Create a new Excel file object.
@@ -41,6 +41,9 @@ use crate::{XlsxColor, XlsxPattern};
 ///     let bold_format = Format::new().set_bold();
 ///     let decimal_format = Format::new().set_num_format("0.000");
 ///     let date_format = Format::new().set_num_format("yyyy-mm-dd");
+///     let merge_format = Format::new()
+///         .set_border(XlsxBorder::Thin)
+///         .set_align(XlsxAlign::Center);
 ///
 ///     // Add a worksheet to the workbook.
 ///     let worksheet = workbook.add_worksheet();
@@ -64,9 +67,12 @@ use crate::{XlsxColor, XlsxPattern};
 ///     // Write a formula.
 ///     worksheet.write_formula_only(5, 0, "=SIN(PI()/4)")?;
 ///
-///     // Write the date .
+///     // Write a date.
 ///     let date = NaiveDate::from_ymd(2023, 1, 25);
 ///     worksheet.write_date(6, 0, date, &date_format)?;
+///
+///     // Write some merged cells.
+///     worksheet.merge_range(7, 0, 7, 1, "Merged cells", &merge_format)?;
 ///
 ///     // Save the file to disk.
 ///     workbook.save("demo.xlsx")?;
@@ -893,8 +899,8 @@ impl Workbook {
             // For a solid fill (pattern == "solid") Excel reverses the role of
             // foreground and background colors, and
             if xf_format.pattern == XlsxPattern::Solid
-                && !xf_format.background_color.is_not_default()
-                && !xf_format.foreground_color.is_not_default()
+                && xf_format.background_color.is_not_default()
+                && xf_format.foreground_color.is_not_default()
             {
                 mem::swap(
                     &mut xf_format.foreground_color,