+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to rust_xlsxwriter will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## In Progress [0.12.0] - 2022-11-XX
+
+### Added
+
+- Added [`worksheet.merge_range()`] method.
+- Added support for Theme colors to [`XlsxColor`]. See also [Working with
+  Colors] in the user guide.
+
+[`XlsxColor`]: https://docs.rs/rust_xlsxwriter/0.11.0/rust_xlsxwriter/enum.XlsxColor.html
+[Working with Colors]: https://rustxlsxwriter.github.io/colors/intro.html
+[`worksheet.merge_range()`]: https://docs.rs/rust_xlsxwriter/latest/rust_xlsxwriter/struct.Worksheet.html#method.merge_range
+
+
 ## [0.11.0] - 2022-11-04
 
 ### Added